move around options
