+++ onmt/train/init.lua
@@ -3,5 +3,6 @@ local train = {}
 train.Checkpoint = require('onmt.train.Checkpoint')
 train.EpochState = require('onmt.train.EpochState')
 train.Optim = require('onmt.train.Optim')
+train.Train = require('onmt.train.Train')
 
 return train
