+++ onmt/utils/Parallel.lua
@@ -19,6 +19,14 @@ local function waitForDevice(dst, src)
    end
 end
 
+function Parallel.declareOpts(cmd)
+  cmd:option('-nparallel', 1, [[When using GPUs, how many batches to execute in parallel.
+                            Note: this will technically change the final batch size to max_batch_size*nparallel.]])
+    cmd:option('-async_parallel', false, [[Use asynchronous parallelism training.]])
+  cmd:option('-async_parallel_minbatch', 1000, [[For async parallel computing, minimal number of batches before being parallel.]])
+  cmd:option('-no_nccl', false, [[Disable usage of nccl in parallel mode.]])  
+end
+
 function Parallel.getCounter()
   local atomic = Parallel._tds.AtomicCounter()
   atomic:inc()
